requirements.txt
flask
gspread
google-auth
google-auth-httplib2
google-auth-oauthlib
requests
streamlit
streamlit-autorefresh
python-dotenv
pandas
